Add files via upload
Task-2 Manual Testing


Q1.You are testing a form that allows users to schedule appointments with a doctor. The
form has the following fields: first name, last name, email, phone number, and appointment
date/time. Some more inputs:
© 

If any field is blank, we should display "All fields are required"
® 
If the email is invalid, we should display "Please enter a valid email"
© 
Ifthe phone number is invalid, we should display "Please enter a valid phone number"

If the appointment date/time is not available, it should display "Please choose another
date/time"
® 
If all fields are correct and the appointment is available, we should schedule the
appointment successfully.

For these requirements write down the correct list of conditions in form of decision table